Stage 6
Connect IRQ line from 6552 to 6502.

Move LCD data pins from 6552 port B to port A, and the control pins from
PA5-7 to PB0-2, to allow us to connect the keyboard clock pin to PB6.

Add keyboard connector. Clock connects via an inverter to CB1, data
connects directly to CB2.

The unused fourth set of inputs on the quad 2-nand gate is used as the
inverter.